requirements.txt
SQLAlchemy == 2.0.44
cors == 1.0.1
Flask-RESTful==0.3.10
python-dotenv==1.0.1
psycopg2-binary==2.9.9
Flask == 3.1.2
Flask-Migrate ==4.1.0